fix(macos): unify mount detection so package status works on macOS
Mount detection was implemented twice: SpawnedMountHandle::is_mounted
(cross-platform, ported in this branch) and manager's check_mount_status
(Linux-only, returned Unknown on macOS). So `packages list` always showed
"Unknown" mount status on macOS even for correctly mounted packages.

Extract one platform-aware is_path_mounted() into system::unmount alongside
unmount_fuse / is_stale_fuse_mount (the platform mount-table module), move
the pure /proc/mounts and mount(8) parsers + their tests there, and have both
SpawnedMountHandle::is_mounted and check_mount_status delegate to it. macOS
now reports real Mounted/NotMounted status instead of Unknown.
